+++ canvas/package.json
@@ -10,6 +10,7 @@
     "dom-to-image": "^2.6.0",
     "file-saver": "^2.0.5",
     "react": "^17.0.2",
+    "react-cookie": "^4.1.1",
     "react-dom": "^17.0.2",
     "react-google-login": "^5.2.2",
     "react-router-dom": "^6.1.1",
